+++ CHANGELOG.md
@@ -5,9 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixes
+- vesicles dataset: failed to get description
+
 ### Added
 - progressbar for downloads (#29)
 
+### Changes (internal)
+- tests with full downloads can be performed with 
+  `pytest -m online`; by default `pytest -m 'not online'` is run,
+  which skips downloading gigabytes of data (PR #18)
+  
 
 ## [0.4.0] - 2018-10-05
 ### Added